move site context to new library structure
