+++ projects/storefrontlib/src/cms-components/misc/site-context-selector/site-context.model.ts
@@ -1,0 +1,4 @@
+export enum SiteContextType {
+  LANGUAGE = 'LANGUAGE',
+  CURRENCY = 'CURRENCY',
+}
