+++ projects/storefrontlib/src/lib/cms-lib/cms-lib.module.ts
@@ -1,5 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { SiteContextSelectorModule } from '../../cms-components/misc';
 import { HamburgerMenuModule, SkipLinkModule } from '../../layout/index';
 import { CartDetailsModule } from '../cart/cart-details/cart-details.module';
 import { CartTotalsModule } from '../cart/cart-totals/cart-totals.module';
@@ -21,7 +22,6 @@ import { NavigationModule } from './navigation/navigation.module';
 import { CmsParagraphModule } from './paragraph/paragraph.module';
 import { ProductCarouselModule } from './product-carousel/product-carousel.module';
 import { SearchBoxModule } from './search-box/search-box.module';
-import { SiteContextSelectorModule } from './site-context-selector/site-context-selector.module';
 
 // import { ProductReferencesModule } from './product-references/product-references.module';
 // import { TabParagraphContainerModule } from './tab-paragraph-container/tab-paragraph-container.module';
@@ -41,9 +41,9 @@ import { SiteContextSelectorModule } from './site-context-selector/site-context-
     ProductCarouselModule,
     SearchBoxModule,
     MiniCartModule,
+    SiteContextSelectorModule,
     // ProductReferencesModule,
     // TabParagraphContainerModule
-    SiteContextSelectorModule,
     AddressBookModule,
     OrderHistoryModule,
     ProductListModule,
