+++ projects/storefrontlib/src/lib/cms-lib/index.ts
@@ -1,33 +1,27 @@
-export * from './cms-lib.module';
-export * from './mini-cart/mini-cart.component';
-export * from './footer-navigation/footer-navigation.component';
-export * from './paragraph/paragraph.component';
-export * from './product-carousel/product-carousel.component';
-export * from './tab-paragraph-container/tab-paragraph-container.component';
-export * from './navigation/navigation.component';
-export * from './category-navigation/category-navigation.component';
-export * from './product-references/product-references.component';
-export * from './link/link.component';
-
-export * from './breadcrumb/breadcrumb.component';
 export * from './banner/banner.component';
-export * from './banner/responsive-banner.component';
-export * from './mini-cart/mini-cart.module';
-export * from './footer-navigation/footer-navigation.module';
-export * from './paragraph/paragraph.module';
-export * from './product-carousel/product-carousel.module';
-export * from './tab-paragraph-container/tab-paragraph-container.module';
-export * from './navigation/navigation.module';
-export * from './category-navigation/category-navigation.module';
-export * from './product-references/product-references.module';
-export * from './link/link.module';
-
-export * from './search-box/search-box.module';
-export * from './search-box/search-box.component';
-export * from './search-box/search-box-component.service';
-
-export * from './breadcrumb/breadcrumb.module';
 export * from './banner/banner.module';
-
-export * from './site-context-selector/site-context-selector.module';
-export * from './site-context-selector/site-context-component.service';
+export * from './banner/responsive-banner.component';
+export * from './breadcrumb/breadcrumb.component';
+export * from './breadcrumb/breadcrumb.module';
+export * from './category-navigation/category-navigation.component';
+export * from './category-navigation/category-navigation.module';
+export * from './cms-lib.module';
+export * from './footer-navigation/footer-navigation.component';
+export * from './footer-navigation/footer-navigation.module';
+export * from './link/link.component';
+export * from './link/link.module';
+export * from './mini-cart/mini-cart.component';
+export * from './mini-cart/mini-cart.module';
+export * from './navigation/navigation.component';
+export * from './navigation/navigation.module';
+export * from './paragraph/paragraph.component';
+export * from './paragraph/paragraph.module';
+export * from './product-carousel/product-carousel.component';
+export * from './product-carousel/product-carousel.module';
+export * from './product-references/product-references.component';
+export * from './product-references/product-references.module';
+export * from './search-box/search-box-component.service';
+export * from './search-box/search-box.component';
+export * from './search-box/search-box.module';
+export * from './tab-paragraph-container/tab-paragraph-container.component';
+export * from './tab-paragraph-container/tab-paragraph-container.module';
