--- projects/storefrontlib/src/lib/cms-lib/site-context-selector/site-context.model.ts
@@ -1,5 +1,0 @@
-
-export enum SiteContextType {
-    LANGUAGE = 'LANGUAGE',
-    CURRENCY = 'CURRENCY',
-  }
